// src/generated/api/public/models/schoolSearchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { SchoolResponse } from "./schoolResponse";

export interface SchoolSearchResponse {
  schools?: SchoolResponse[];
  page?: number;
  pageSize?: number;
  totalCount?: number;
}