// src/generated/api/public/models/schoolResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */

export interface SchoolResponse {
  id?: string;
  neisSchoolCode?: string;
  educationOfficeCode?: string;
  name?: string;
  address?: string;
  region?: string;
}